# exclude labels formatting for every operation in LOGICAL_TIMELINE_SIZE