implements solution suggested by @hlinnaka here: https://github.com/zenithdb/zenith/issues/622#issuecomment-921904499

I do not see these calls anywhere near the top in perf, @hlinnaka could you please verify?

resolves #622