@hlinnaka 2021-09-21
> Running perf on that now on my laptop, `increase_current_logical_size` accounted for about 8% of the pageserver CPU time in that test. That's noticable, but less than I saw earlier; I think I had some other quick & dirty optimizations in my working directory that made `increase_current_logical_size` stick out even more.

I remembered one little thing I was experimenting with then, and put that in a PR: https://github.com/zenithdb/zenith/pull/641.