@hlinnaka 2021-09-21
Let's push this only after the new test in https://github.com/zenithdb/zenith/pull/640. Then we'll get a baseline run without this, and we can see the improvement when this is pushed.