@LizardWizzard 2021-09-22
Before:
```
test_bulk_insert.insert: 65.926 s
test_bulk_insert.pageserver_writes: 1,426 MB
test_bulk_insert.size: 914 MB
test_pgbench.init: 8.962 s
test_pgbench.pageserver_writes: 466 MB
test_pgbench.5000_xacts: 10.413 s
test_pgbench.size: 163 MB
```
After:
```
test_bulk_insert.insert: 63.311 s
test_bulk_insert.pageserver_writes: 1,482 MB
test_bulk_insert.size: 914 MB
test_pgbench.init: 6.032 s
test_pgbench.pageserver_writes: 466 MB
test_pgbench.5000_xacts: 9.173 s
test_pgbench.size: 163 MB
```
cc @hlinnaka 

It is strange that benchmark after has more pageserver writes, it shouln't be changed, I think it is just a bit imprecise or bad timing with checkpoint etc.