@hlinnaka 2021-09-22
> It is strange that benchmark after has more pageserver writes, it shouln't be changed, I think it is just a bit imprecise or bad timing with checkpoint etc.

Yeah, it fluctuates a bit, that's normal